Update patch adapter
First, setup ability to map classes in patch
to newclasses that may be created for them to
host the stolen fields and methods.

Second, for method invoke statements that have multiple targets
loop those targets and if any match a method that we moved,
construct a guard that maps the runtime type of the class to
the type we moved a method OUT of.