Fix patch adapter old to new matching strategy
Previously adapter used a map of old method refs
to new ones to fix method refs, however in the
case where a child class relies on a parent def of
a method, the method ref there will refer to the
child and just resolve to parent at runtime.
therefore we were not fixing those refs, when we
should be.
to fix: use just a list of new methods and if any
target of a call is in the newmethods list, use that.