Update patch adapter and differ
Differ updated to diff on all of the original
classes which are actually mapped to a redef,
just in case some extra stuff ends up in the
directory that it is supposed to read the
diff resources from.

Update adapter to construct a call to replace
the method ref in the else condition of the guard.
However, this is not really type safe, will fail on
some inheritance schemas