Fix class renaming bug
asm classwriter will need to access sootclasses
that have been renamed in the case where there is
common super types. For some reason these classname
to type pairings were not explicitly in the scene's
nametoclass map, and was causing an error. now we
explicitly add the entry to the map