patch adapter enhancements
make added methods that get stolen into hosts
protected not private so that they can be accessed
in original class.
also
make stolen fields static so that any instance of newclass host
can share their state. this is needed so that any method in the
original class can modify a field and see the lasting effects of this
just like as if it belonged to that class. Otherwise would have to pass
an instance of the host class around, and that would be not feasible